[python-sdk] moved url constants
