+++ razorpay/constants/__init__.py
@@ -1,7 +1,9 @@
 from .http_status_code import HTTP_STATUS_CODE
 from .error_code import ERROR_CODE
+from .url import URL
 
 __all__ = [
         'HTTP_STATUS_CODE',
         'ERROR_CODE',
+        'URL',
 ]
