+++ razorpay/constants/url.py
@@ -1,4 +1,5 @@
 class URL(object):
+    BASE_URL = 'https://api.razorpay.com/v1'
     ORDER_URL = "/orders"
     INVOICE_URL = "/invoices"
     PAYMENTS_URL = "/payments"
